01_model_and_data/02_model_and_data_care/airflow_extract_sql/requirements.txt
# requirements.txt

ccxt
apache-airflow-providers-docker
docker
python-dotenv